Inference/requirements.txt
catboost
uvicorn[standard]
fastapi
scikit-learn
pyyaml